let's pushhhh

+++ ft_strchr.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zmourid <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: zmourid <zmourid@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/05 15:52:37 by zmourid           #+#    #+#             */
-/*   Updated: 2023/11/07 01:38:00 by zmourid          ###   ########.fr       */
+/*   Updated: 2023/11/18 01:04:43 by zmourid          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
 #include "libft.h"
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	
 	if ((char)c == '\0')
 		return ((char *)s + ft_strlen(s));
 	while (s[i])
